Move dig() to utils.dns
