+++ data/hooks/diagnosis/12-dnsrecords.py
@@ -8,11 +8,10 @@ from publicsuffixlist import PublicSuffixList
 
 from moulinette.utils.process import check_output
 
-from yunohost.utils.network import dig
+from yunohost.utils.dns import dig, YNH_DYNDNS_DOMAINS
 from yunohost.diagnosis import Diagnoser
 from yunohost.domain import domain_list, _build_dns_conf, _get_maindomain
 
-YNH_DYNDNS_DOMAINS = ["nohost.me", "noho.st", "ynh.fr"]
 SPECIAL_USE_TLDS = ["local", "localhost", "onion", "test"]
